Added most test
+ User
+ SmartPhone
+ Cover
- Order

+++ SmartPhoneAndCoverWebshop/SmartPhoneShop.Core/ApplicationService/impl/CoverService.cs
@@ -17,41 +17,45 @@ namespace SmartPhoneShop.Core.ApplicationService.impl
 
         public Cover CreateCover(Cover CreatedCover)
         {
-            if (CreatedCover.Name == null)
+            if (string.IsNullOrEmpty(CreatedCover.Name))
             {
                 throw new InvalidDataException("Must have a product name");
             }
 
-            if (CreatedCover.Color == null)
+            if (string.IsNullOrEmpty(CreatedCover.Color))
             {
                 throw new InvalidDataException("Everything has a color, so it must have a color");
             }
 
-            if (CreatedCover.Material == null)
+            if (string.IsNullOrEmpty(CreatedCover.Material))
             {
                 throw new InvalidDataException("Must be made with a material");
             }
 
-            if (CreatedCover.TypeOfBrand == null)
+            if (string.IsNullOrEmpty(CreatedCover.TypeOfBrand))
             {
                 throw new InvalidDataException("Must have a brand for this cover");
             }
 
-
-            if (CreatedCover.TypeOfModel == null)
+            if (string.IsNullOrEmpty(CreatedCover.TypeOfModel))
             {
                 throw new InvalidDataException("Must have a Phone Model to this Cover");
             }
 
-            if (CreatedCover.Price.Equals(null))
+            if (CreatedCover.Price <= 0)
             {
                 throw new InvalidDataException("Must have a price");
             }
+
             return _coverRepository.CreateCover(CreatedCover);
         }
 
         public Cover DeleteCover(int id)
         {
+            if (_coverRepository.DeleteCover(id) == null)
+            {
+                throw new InvalidDataException("No Cover with id: " + id + " exist");
+            }
             return _coverRepository.DeleteCover(id);
         }
 
@@ -71,6 +75,34 @@ namespace SmartPhoneShop.Core.ApplicationService.impl
 
         public Cover UpdateCover(Cover UpdateCover)
         {
+            if (string.IsNullOrEmpty(UpdateCover.Name))
+            {
+                throw new InvalidDataException("Must have a name");
+            }
+            if (string.IsNullOrEmpty(UpdateCover.Color))
+            {
+                throw new InvalidDataException("Everything has a color, so it must have a color");
+            }
+
+            if (string.IsNullOrEmpty(UpdateCover.Material))
+            {
+                throw new InvalidDataException("Must be made with a material");
+            }
+
+            if (string.IsNullOrEmpty(UpdateCover.TypeOfBrand))
+            {
+                throw new InvalidDataException("Must have a brand for this cover");
+            }
+
+            if (string.IsNullOrEmpty(UpdateCover.TypeOfModel))
+            {
+                throw new InvalidDataException("Must have a Phone Model to this Cover");
+            }
+
+            if (UpdateCover.Price <= 0)
+            {
+                throw new InvalidDataException("Must have a price");
+            }
             return _coverRepository.UpdateCover(UpdateCover);
         }
     }
