+++ SmartPhoneAndCoverWebshop/SmartPhoneShop.Core/ApplicationService/impl/SmartPhoneService.cs
@@ -19,37 +19,37 @@ namespace SmartPhoneShop.Core.ApplicationService.impl
 
         public SmartPhone CreateSmartPhone(SmartPhone CreatedSmartPhone)
         {
-            if (CreatedSmartPhone.Name == null)
+            if (string.IsNullOrEmpty(CreatedSmartPhone.Name))
             {
                 throw new InvalidDataException("Must have a product name");
             }
 
-            if (CreatedSmartPhone.Price.Equals(null))
+            if (CreatedSmartPhone.Price <= 0)
             {
                 throw new InvalidDataException("Must have a price");
             }
 
-            if (CreatedSmartPhone.Camera == null)
+            if (string.IsNullOrEmpty(CreatedSmartPhone.Camera))
             {
                 throw new InvalidDataException("Phones always have a camera");
             }
 
-            if (CreatedSmartPhone.CpuType == null)
+            if (string.IsNullOrEmpty(CreatedSmartPhone.CpuType))
             {
                 throw new InvalidDataException("Must have a CPU");
             }
 
-            if (CreatedSmartPhone.Memory.Equals(null) || CreatedSmartPhone.Memory <= 0)
+            if (CreatedSmartPhone.Memory <= 0)
             {
                 throw new InvalidDataException("Must have a certain amount of memory");
             }
 
-            if (CreatedSmartPhone.OS == null)
+            if (string.IsNullOrEmpty(CreatedSmartPhone.OS))
             {
                 throw new InvalidDataException("Must have a OS");
             }
 
-            if(CreatedSmartPhone.Screen.Equals(null) || CreatedSmartPhone.Screen <= 0)
+            if(CreatedSmartPhone.Screen <= 0)
             {
                 throw new InvalidDataException("Must have a screen size");
             }
@@ -58,6 +58,10 @@ namespace SmartPhoneShop.Core.ApplicationService.impl
 
         public SmartPhone DeleteSmartPhone(int id)
         {
+            if (_smartPhoneRepository.DeleteSmartPhone(id) == null)
+            {
+                throw new InvalidDataException("No SmartPhone with id: " + id + " exist");
+            }
             var phone = _smartPhoneRepository.DeleteSmartPhone(id);
             return phone;
         }
@@ -74,6 +78,40 @@ namespace SmartPhoneShop.Core.ApplicationService.impl
 
         public SmartPhone UpdateSmartPhone(SmartPhone UpdateSmartPhone)
         {
+            if (string.IsNullOrEmpty(UpdateSmartPhone.Name))
+            {
+                throw new InvalidDataException("Must have a product name");
+            }
+
+            if (UpdateSmartPhone.Price <= 0)
+            {
+                throw new InvalidDataException("Must have a price");
+            }
+
+            if (string.IsNullOrEmpty(UpdateSmartPhone.Camera))
+            {
+                throw new InvalidDataException("Phones always have a camera");
+            }
+
+            if (string.IsNullOrEmpty(UpdateSmartPhone.CpuType))
+            {
+                throw new InvalidDataException("Must have a CPU");
+            }
+
+            if (UpdateSmartPhone.Memory <= 0)
+            {
+                throw new InvalidDataException("Must have a certain amount of memory");
+            }
+
+            if (string.IsNullOrEmpty(UpdateSmartPhone.OS))
+            {
+                throw new InvalidDataException("Must have a OS");
+            }
+
+            if(UpdateSmartPhone.Screen <= 0)
+            {
+                throw new InvalidDataException("Must have a screen size");
+            }
             return _smartPhoneRepository.UpdateSmartPhone(UpdateSmartPhone);
 
         }
